requirements.txt
asgiref
Django
django-cors-headers
django-environ
django-mysql
djangorestframework
django-oauth-toolkit
pymysql
graphene-django
graphene-mongo
mongoengine
mongomock